# docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact:
    email: admin@kavling-mentaya.com
    name: API Support
  description: REST API backend untuk sistem manajemen penjualan kavling tanah. Semua
    endpoint (kecuali /auth/login) memerlukan Bearer Token JWT di header Authorization.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Kavling Mentaya API
  version: "1.0"
paths:
...
  /laporan/export-excel:
    get:
      description: Mengunduh laporan arus kas dalam format .xlsx dengan filter rentang
        tanggal
      parameters:
      - description: Dari tanggal (YYYY-MM-DD)
        in: query
        name: dari
        type: string
      - description: Sampai tanggal (YYYY-MM-DD)
        in: query
        name: sampai
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Export laporan arus kas ke Excel
      tags:
      - Laporan
  /laporan/export-excel-pembayaran:
    get:
      description: Mengunduh laporan riwayat pembayaran cicilan dalam format .xlsx
      parameters:
      - description: Dari tanggal (YYYY-MM-DD)
        in: query
        name: dari
        type: string
      - description: Sampai tanggal (YYYY-MM-DD)
        in: query
        name: sampai
        type: string
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: OK
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Export laporan pembayaran cicilan ke Excel
      tags:
      - Laporan
  /laporan/per-customer:
    get:
      description: Mengambil riwayat pembayaran/transaksi untuk customer tertentu
        dalam rentang tanggal
      parameters:
      - description: Customer ID
        in: query
        name: id_customer
        required: true
        type: integer
      - description: Dari tanggal (YYYY-MM-DD)
        in: query
        name: dari
        type: string
      - description: Sampai tanggal (YYYY-MM-DD)
        in: query
        name: sampai
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Laporan per customer
      tags:
      - Laporan
  /laporan/per-kategori:
    get:
      description: Mengambil transaksi arus kas berdasarkan kategori tertentu dalam
        rentang tanggal
      parameters:
      - description: 'Nama kategori (contoh: Bayar Angsuran)'
        in: query
        name: kategori
        required: true
        type: string
      - description: Dari tanggal (YYYY-MM-DD)
        in: query
        name: dari
        type: string
      - description: Sampai tanggal (YYYY-MM-DD)
        in: query
        name: sampai
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Laporan per kategori
      tags:
      - Laporan
  /laporan/umum:
    get:
      description: Mengambil laporan arus kas dengan filter rentang tanggal beserta
        total pemasukan, pengeluaran, dan saldo
      parameters:
      - description: Dari tanggal (YYYY-MM-DD)
        in: query
        name: dari
        type: string
      - description: Sampai tanggal (YYYY-MM-DD)
        in: query
        name: sampai
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: Laporan arus kas umum
      tags:
      - Laporan
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}. Token diperoleh
      dari endpoint POST /api/auth/login.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"